+++ demos/chinook/src/main/java/is/codion/demos/chinook/model/InvoiceModel.java
@@ -18,9 +18,7 @@
  */
 package is.codion.demos.chinook.model;
 
-import is.codion.demos.chinook.domain.api.Chinook.InvoiceLine;
 import is.codion.framework.db.EntityConnectionProvider;
-import is.codion.framework.model.ForeignKeyModelLink;
 import is.codion.swing.framework.model.SwingEntityModel;
 
 import javax.swing.SwingUtilities;
@@ -33,7 +31,7 @@ public final class InvoiceModel extends SwingEntityModel {
 		InvoiceLineEditModel invoiceLineEditModel = new InvoiceLineEditModel(connectionProvider);
 
 		SwingEntityModel invoiceLineModel = new SwingEntityModel(invoiceLineEditModel);
-		detailModels().add(ForeignKeyModelLink.builder(invoiceLineModel, InvoiceLine.INVOICE_FK)
+		detailModels().add(link(invoiceLineModel)
 						// Prevents accidentally adding a new invoice line to the previously selected invoice,
 						// since the selected foreign key value persists when the master selection is cleared by default.
 						.clearValueOnEmptySelection(true)