+++ framework/model/src/test/java/is/codion/framework/model/DefaultEntityModelTest.java
@@ -35,7 +35,7 @@ public class DefaultEntityModelTest extends AbstractEntityModelTest<
 	protected TestEntityModel createDepartmentModel() {
 		TestEntityModel deptModel = new TestEntityModel(new TestEntityEditModel(Department.TYPE, connectionProvider()));
 		TestEntityModel empModel = new TestEntityModel(new TestEntityEditModel(Employee.TYPE, connectionProvider()));
-		deptModel.detailModels().add(ModelLink.builder(empModel)
+		deptModel.detailModels().add(deptModel.link(empModel)
 						.active(true)
 						.build());