+++ changelog.md
@@ -40,6 +40,7 @@ Codion Change Log
 - EntityModel.DetailModels.link() replaced with active().
 - ModelLink, type parameters removed.
 - EntityModel, M type parameter removed.
+- EntityModel.link() added, now throws exception in case multiple fitting foreign keys are found.
 ### is.codion.swing.framework.model
 - DefaultSwingForeignKeyConditionModel now links search and combo box models to their respective operands.
 ### is.codion.swing.framework.ui